requirements.txt
alpaca-py==0.0.14
requests==2.28.1
python-binance==1.0.18
pandas==1.5.3
matplotlib==3.7.1
numpy==1.23.5